requirements.txt
colorama
requests
discord.py-self==1.9.2